cast to int32

+++ nilearn/plotting/img_plotting.py
@@ -808,7 +808,7 @@ def _plot_roi_contours(display, roi_img, cmap, alpha, linewidths):
         if label == 0:
             continue
         data = roi_data == label
-        data = data.astype(int)
+        data = data.astype(np.int32)
         img = new_img_like(roi_img, data, affine=roi_img.affine)
         display.add_contours(
             img,
